api_gateway/requirements.txt
# Core Flask dependencies
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8

# Production Server
gunicorn==23.0.0

# Database & ORM
SQLAlchemy==2.0.28
pymysql==1.1.0

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
requests==2.32.3
urllib3==2.3.0
idna==3.10
charset-normalizer==3.4.1
certifi==2025.1.31

# Misc Utilities
blinker==1.9.0
importlib_metadata==8.6.1
MarkupSafe==3.0.2
packaging==24.2
zipp==3.21.0
